[ENH]: Make foundation sync collections use an uber function
